✨ feat: improve coin-base build result structure
