+++ package.json
@@ -19,8 +19,8 @@
   "workspaces": [
     "packages/*"
   ],
-  "dependencies": {
-    "bignumber.js": "^9.1.2"
+  "peerDependencies": {
+    "bignumber.js": ">=9.0.0"
   },
   "devDependencies": {
     "dotenv": "^16.4.5",
