+++ packages/coin-base/src/derive/index.ts
@@ -1,2 +1,3 @@
 export { getEd25519DerivedPrivateKey, getRandomEd25519PrivateKey } from './ed25519';
 export * from './ecdsa';
+export * from './base-chain-wallet'
