+++ packages/coin-base/src/model.ts
@@ -1,0 +1,4 @@
+import 'reflect-metadata';
+
+export * from './types';
+export * from './enums';
