+++ packages/coin-ethereum/package.json
@@ -39,8 +39,10 @@
     "format": "prettier --write ."
   },
   "dependencies": {
-    "@delandlabs/coin-base": "*",
     "ethers": "^6.13.5"
+  },
+  "peerDependencies": {
+    "@delandlabs/coin-base": "*"
   },
   "devDependencies": {
     "@eslint/js": "^9.16.0",
