+++ packages/coin-ethereum/src/chains.ts
@@ -1,4 +1,4 @@
-import { ChainInfo, ChainId, Chain, ChainNetwork, Ecosystem, WalletSignatureSchema } from "@delandlabs/coin-base";
+import { ChainInfo, ChainId, Chain, ChainNetwork, Ecosystem, WalletSignatureSchema } from "@delandlabs/coin-base/model";
 
 export const Ethereum: ChainInfo = {
   chainId: new ChainId(Chain.Ethereum, ChainNetwork.EvmMainNet),
