+++ packages/coin-kaspa/src/chain-wallet/chain.ts
@@ -1,4 +1,4 @@
-import { ChainInfo, ChainId, Chain, ChainNetwork, Ecosystem, WalletSignatureSchema } from '@delandlabs/coin-base';
+import { ChainInfo, ChainId, Chain, ChainNetwork, Ecosystem, WalletSignatureSchema } from '@delandlabs/coin-base/model';
 
 export const Kaspa: ChainInfo = {
   chainId: new ChainId(Chain.Kaspa, ChainNetwork.KaspaMainNet),
