+++ packages/coin-ton/src/chain-wallet/chain.ts
@@ -1,4 +1,4 @@
-import { ChainInfo, ChainId, Chain, ChainNetwork, Ecosystem, WalletSignatureSchema } from '@delandlabs/coin-base';
+import { ChainInfo, ChainId, Chain, ChainNetwork, Ecosystem, WalletSignatureSchema } from '@delandlabs/coin-base/model';
 import { getHttpEndpoint } from '@orbs-network/ton-access';
 import TonWeb from 'tonweb';
